Python_Code/api/requirements.txt
pandas==2.0.3
python_dotenv==1.0.1
mlxtend==0.23.1
pinecone
pandas==2.0.3
python-dotenv>=1.0.0
runpod==1.7.1
pinecone==5.3.1
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-ollama>=0.0.1
langchain-groq>=0.0.1
pydantic>=2.0.0